src/rApp/requirements.txt
fastapi==0.119.1
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5
uvicorn==0.38.0
